feat: Allow for attaching access policies to applications (#4)
Allows us to attach access polices to applications.

The policy must already exist without cloudflare and modifications to
the applications policy through the UI will be replaced to respect the
expected result from the application definition.

The policy precedence will be defined as the order the polices appear in
the resource.